@@ -1,44 +1,46 @@
 {
   "name": "tailwind-scrollbar-hide",
-  "version": "1.1.7",
+  "version": "1.2.0",
   "description": "tailwindcss plugin for hide scrollbar",
-  "main": "src/index.js",
-  "scripts": {
-    "dev": "vite ./playground"
+  "keywords": [
+    "tailwindcss",
+    "scrollbar",
+    "scrollbar-hide"
+  ],
+  "homepage": "https://github.com/reslear/tailwind-scrollbar-hide#readme",
+  "bugs": {
+    "url": "https://github.com/reslear/tailwind-scrollbar-hide/issues"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/reslear/tailwind-scrollbar-hide.git"
   },
-  "files": [
-    "src"
-  ],
-  "keywords": [
-    "tailwindcss",
-    "scrollbar"
-  ],
-  "author": "reslear",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/reslear/tailwind-scrollbar-hide/issues"
+  "author": "reslear",
+  "sideEffects": false,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
   },
-  "homepage": "https://github.com/reslear/tailwind-scrollbar-hide#readme",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "prepublishOnly": "pnpm build"
+  },
   "devDependencies": {
-    "@semantic-release/changelog": "6.0.1",
-    "@semantic-release/git": "10.0.1",
-    "@types/node": "14.14.25",
-    "@vitejs/plugin-vue": "1.1.4",
-    "@vue/compiler-sfc": "3.0.5",
-    "@vuedx/typecheck": "0.6.3",
-    "@vuedx/typescript-plugin-vue": "0.6.3",
-    "autoprefixer": "10.2.4",
-    "conventional-changelog-conventionalcommits": "^4.6.1",
-    "postcss": "8.2.5",
-    "semantic-release": "18.0.0",
-    "tailwindcss": "2.0.2",
-    "typescript": "4.1.3",
-    "vite": "2.0.0-beta.65",
-    "vue": "3.0.5"
+    "@types/node": "^22.10.2",
+    "tailwindcss": "3.4.17",
+    "typescript": "5.7.2"
+  },
+  "peerDependencies": {
+    "tailwindcss": ">=3.0.0 || >= 4.0.0 || >= 4.0.0-beta.8 || >= 4.0.0-alpha.20"
   },
   "publishConfig": {
     "access": "public"
