@@ -1,20 +1,20 @@
 {
-  "name": "vite-vue-typescript-starter",
+  "name": "playground",
+  "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vuedx-typecheck . && vite build",
-    "serve": "vite preview"
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "vue": "3.0.5"
+    "@tailwindcss/vite": "4.0.0-beta.8",
+    "tailwind-scrollbar-hide": "workspace:*",
+    "tailwindcss": "4.0.0-beta.8"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "1.1.4",
-    "@vue/compiler-sfc": "3.0.5",
-    "@vuedx/typecheck": "0.6.3",
-    "@vuedx/typescript-plugin-vue": "0.6.3",
-    "typescript": "4.1.3",
-    "vite": "2.0.0-beta.65"
+    "typescript": "5.7.2",
+    "vite": "6.0.3"
   }
 }
