@@ -1,6 +1,1 @@
-import { createApp } from 'vue'
-
-import App from './App.vue'
-import './index.css'
-
-createApp(App).mount('#app')
+import './style.css'
